split monolith
