+++ .vscode/launch.json
@@ -1,0 +1,31 @@
+{
+	// Use IntelliSense to learn about possible attributes.
+	// Hover to view descriptions of existing attributes.
+	// For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+	"version": "0.2.0",
+	"configurations": [
+		{
+			"type": "lldb",
+			"request": "launch",
+			"name": "Debug executable 'karthograph'",
+			"cargo": {
+				"args": [
+					"build",
+					"--bin=karthograph",
+					"--package=karthograph"
+				],
+				"filter": {
+					"name": "karthograph",
+					"kind": "bin"
+				}
+			},
+			"args": [],
+			"env": {
+				"LD_LIBRARY_PATH": "${env:LD_LIBRARY_PATH}:${workspaceFolder}/target/debug/deps/:${env:HOME}/.rustup/toolchains/nightly-x86_64-unknown-linux-gnu/lib",
+				"CARGO_MANIFEST_DIR": "${workspaceFolder}"
+			},
+			"cwd": "${workspaceFolder}/../..",
+			"relativePathBase": "${workspaceFolder}/../.."
+		},
+	]
+}
