+++ karthograph_client/Cargo.toml
@@ -6,3 +6,5 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+bevy = { version = "0.5.0", features = ["dynamic"] }
+karthograph_core = {path = "../karthograph_core"}
