+++ karthograph_client/assets/third_party_assets.md
