+++ karthograph_client/assets/village.png
