+++ karthograph_core/src/asset_management.rs
