+++ karthograph_core/src/grid.rs
