+++ karthograph_core/src/seasons.rs
