+++ karthograph_core/src/shape.rs
