+++ karthograph_core/src/util.rs
